Call gbvar->ComputeShear to set shear parameters
ComputeShear call in the time integrator sets shear parameters of
the *selected* BoundaryVariable instances that are enrolled in
'bvars_main_int' vector.
Since gravity is not contained in the time integrator (it has a
seperate TaskList), we may not want to enroll 'gbvar' to the
'bvars_main_int'.
In that case, we need to explicitly call ComputeShear to set up
shear parameters of the gbvar instance.