Interpolate Phi within nearest integer times
The 'phase shift' method has a inherent flaw: the potential is
accurate only when the sheared distance is a integer multiple of
\Delta y.
For arbitrary non-integer times, we linearly intepolate Phi in time
from nearest integer times to get second-order convergence.
This requires solving Poisson equation twice; we will use roll&unroll
method in the future.