Remove redundant ComputeShear call in FFTGravitySolverTaskList
ComputeShear is already called in TimeIntegratorTaskList which come earlier
than FFTGravitySolverTaskList. Reusing the shear parameters that are already
set by ComputeShear call in the TimeIntegratorTaskList guarentees that the
shear parameters for gravity and hydro are identical.

Also, this commit moves the initial gravity solve in the mesh
constructor below the ComputeShear call, consistent with the main
loop.

See issue #20 for related discussion.