Solve initial Poisson after hydro boundary comms.
Solve Poisson equation at the mesh initialization step after
the hydro boundary communications. In this way, the density
at the ghost zones in the y boundaries are correctly filled.
We need this because RollUnroll requires +-1 neighboring cells
to do the PLM interpolation.

Unlike the y boundaries, the density at the ghost zones in the
x boundaries (shearing boundaries) is not filled in the
initialization step. This should not cause any problem, because
they are eventually filled in the main integration loop.
Be aware that when <output>/ghost_zones=true, the density at the
x boundaries in the first output file is not consistent with
the shearing BC.