Implement fractional shift
For fractional shift, we use RemapFluxPlm or RemapFluxPpm function
of OrbitalAdvection class, depending on the order of spatial
reconstruction (xorder).

To use those function, we make y axis fast (to minimize
cache miss). Note that both the current orbital advection module
and the Athena RemapVar functions use (k,i,j) storage order, but
it seems that (i,k,j) is more cache efficient considering the loop
order. Need to test this in the future.

Unlike the input density which is already ghost-zone-filled, the
rolled potential lives only in the active zones and therefore need
communication before unroll. This is also implemented in this commit.