Revert "Compute correct shear at initial time"
This reverts commit 3f0d61cb14b7dd75e422e33b1533fa4987e65a84.
The original motivation of the above commit was to fill the ghost
zone potential of the first vtk, especially for one-time poisson
solve problem that can have weird dt.

However, this can lead to wrong shear info at the first timestep,
ncycle=0, so should be reverted.

Note that the ghost zone potential of the initial vtk dump is only
for checking and do not affect the main integration (gravity is solved
again at ncycle=0 in the main loop).