Revise confusing variable names
Previous ppm->weight was not very meaningful, as it was a sum
of the cloud-in-cell weights of all particles that fall in a given
cell. Moreover, it was later multiplied by mass, so in fact it
represents mass density (except for StarParticle where weight becomes
number density and seperate ppm->density contains mass density, which
is even more confusing).

This commit retires the number density; FindLocalDensity calculates
mass and momentum density by default. Variable names are changed
accordingly. Number density field may be added in meshaux later.