Redesign ghost particle communication
There was a problem with the previous design, in which the particles
in the overlap region are appended to the send buffer and are sent
together with active particles: when a particle crosses the MeshBlock
boundary, the receiving MeshBlock has not yet owns that particle, so
it cannot broadcast the particle to its neighbors. That is, the
ghost particle communication should be done *after* active particle
communication is completed.